move output_capture to utils
