+++ lib/ramble/ramble/appkit.py
@@ -30,4 +30,4 @@ from ramble.util.logger import logger as tty
 
 from ramble.util.file_util import get_file_path
 
-from ramble.schema.types import OUTPUT_CAPTURE
+from ramble.util.output_capture import OUTPUT_CAPTURE
