+++ lib/ramble/ramble/schema/types.py
@@ -8,33 +8,6 @@
 
 """Base types for building schema files from"""
 
-from enum import Enum
-
-
-class OUTPUT_CAPTURE(Enum):
-    STDOUT = 0
-    STDERR = 1
-    ALL = 2
-    DEFAULT = 2
-
-
-class output_mapper:
-    def __init__(self):
-        self.map = {
-            OUTPUT_CAPTURE.STDOUT: ">>",
-            OUTPUT_CAPTURE.STDERR: "2>>",
-            OUTPUT_CAPTURE.ALL: ">>",
-        }
-        self.SUFFIX = "2>&1"
-
-    def generate_out_string(self, out_log, output_operator):
-        redirect_str = f' {self.map.get(output_operator, output_operator)} "{out_log}"'
-        if output_operator is OUTPUT_CAPTURE.ALL:
-            redirect_str += f" {self.SUFFIX}"
-
-        return redirect_str
-
-
 string_or_num_list = [{"type": "string"}, {"type": "number"}]
 
 string_or_num = {"anyOf": string_or_num_list}
